LayoutTests/imported/w3c/web-platform-tests/content-security-policy/reporting-api/reporting-api-report-only-sends-reports-on-violation.https.sub-expected.txt



PASS Test that image does load
PASS Event is fired
PASS Violation report status OK.
